pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-chaser-mcp"
version = "0.2.2"
description = "Model Context Protocol server for academic paper search across Semantic Scholar, CORE, arXiv, OpenAlex, SerpApi Google Scholar, Crossref, Unpaywall, ECOS species dossiers, and Federal Register / CFR regulatory text"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["mcp", "semantic-scholar", "core", "arxiv", "openalex", "academic", "papers", "research", "federal-register", "ecos", "regulatory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "anyio>=4.0.0",
    "defusedxml>=0.7.1",
    "fastmcp>=3.1.0",
    "mcp>=1.24.0",
    "httpx>=0.25.0",
    "markitdown[pdf]>=0.1.5",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
ai = [
    "langchain>=1.2.13",
    "langgraph>=1.1.3",
    "langgraph-checkpoint>=4.0.1",
    "numpy>=2.2.6,<2.4.0; python_version < '3.11'",
    "numpy>=2.4.3; python_version >= '3.11'",
]
openai = [
    "langchain-openai>=1.1.11",
    "openai>=1.109.0",
]
huggingface = [
    "langchain-openai>=1.1.11",
    "openai>=1.109.0",
]
nvidia = [
    "langchain-nvidia-ai-endpoints>=0.4.0",
]
anthropic = [
    "anthropic>=0.86.0",
    "langchain-anthropic>=1.4.0",
]
google = [
    "google-genai>=1.69.0",
    "langchain-google-genai>=4.2.1",
]
mistral = [
    "langchain-mistralai>=0.2.5",
    "mistralai>=1.9.11",
]
eval-foundry = [
    "azure-ai-projects>=2.0.1",
    "azure-identity>=1.25.3",
]
eval-huggingface = [
    "huggingface_hub>=1.9.0",
]
eval = [
    "paper-chaser-mcp[eval-foundry,eval-huggingface]",
]
ai-faiss = [
    "faiss-cpu>=1.13.2",
    "langchain-community>=0.4.1",
]
dev = [
    "bandit>=1.8.0",
    "build>=1.4.0",
    "bumpver>=2025.1131",
    "mypy>=1.17.0",
    "pip-audit>=2.10.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.12.0",
    "shellcheck-py>=0.11.0.1",
    "types-defusedxml>=0.7.0",
]
all = [
    "paper-chaser-mcp[ai,openai,huggingface,nvidia,anthropic,google,mistral,eval,dev]",
]

[project.scripts]
paper-chaser-mcp = "paper_chaser_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/joshuasundance-swca/paper-chaser-mcp"
Repository = "https://github.com/joshuasundance-swca/paper-chaser-mcp"
Issues = "https://github.com/joshuasundance-swca/paper-chaser-mcp/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["paper_chaser_mcp*"]

[tool.bumpver]
current_version = "v0.2.2"
version_pattern = "vMAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
commit = false
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{pep440_version}"',
]
"paper_chaser_mcp/__init__.py" = [
    '__version__ = "{pep440_version}"',
]
"server.json" = [
    '"version": "{pep440_version}",',
    '"identifier": "ghcr.io/joshuasundance-swca/paper-chaser-mcp:{pep440_version}",',
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-ra --strict-config --strict-markers"
xfail_strict = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
]

[tool.mypy]
python_version = "3.10"
files = ["paper_chaser_mcp", "tests"]
check_untyped_defs = true

# Optional agentic extras (ai, ai-faiss) are not installed in CI base runs.
# These imports are guarded at runtime, so silence mypy when stubs are absent.
[[tool.mypy.overrides]]
module = [
    "anthropic.*",
    "google.*",
    "openai.*",
    "langchain.*",
    "langchain_anthropic.*",
    "langchain_google_genai.*",
    "langchain_nvidia_ai_endpoints.*",
    "langchain_mistralai.*",
    "langchain_openai.*",
    "langchain_community.*",
    "langchain_core.*",
    "langgraph.*",
    "faiss.*",
    "mistralai.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.coverage.run]
source = ["paper_chaser_mcp"]

[tool.coverage.report]
fail_under = 85
show_missing = true
skip_covered = true
